Gave logout route some error handling

+++ app/controllers/sessions_controller.rb
@@ -12,7 +12,11 @@ class SessionsController < ApplicationController
     end
 
     def destroy
-        session.delete(:user_id)
-        head :no_content
+        if session[:user_id]
+            session.delete(:user_id)
+            head :no_content
+        else
+            render json: {errors: ["You can not logout without being logged in"]}, status: :unauthorized
+        end
     end
 end
